@@ -13,49 +13,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.2",
-    "@angular/compiler": "^4.0.2",
-    "@angular/core": "^4.0.2",
-    "@angular/forms": "^4.0.2",
-    "@angular/http": "^4.0.2",
-    "@angular/platform-browser": "^4.0.2",
-    "@angular/platform-browser-dynamic": "^4.0.2",
-    "@angular/router": "^4.0.2",
+    "@angular/common": "^4.1.3",
+    "@angular/compiler": "^4.1.3",
+    "@angular/core": "^4.1.3",
+    "@angular/forms": "^4.1.3",
+    "@angular/http": "^4.1.3",
+    "@angular/platform-browser": "^4.1.3",
+    "@angular/platform-browser-dynamic": "^4.1.3",
+    "@angular/router": "^4.1.3",
     "angular2-jwt": "0.2.2",
     "core-js": "^2.4.1",
     "extract-text-webpack-plugin": "2.0.0-beta.4",
     "font-awesome": "^4.7.0",
     "moment": "^2.14.1",
-    "ng2-bootstrap": "1.6.3",
+    "ngx-bootstrap": "1.7.1",
     "ng2-fontawesome": "0.0.6",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "5.1.0",
-    "zone.js": "^0.8.4"
+    "reflect-metadata": "^0.1.10",
+    "rxjs": "5.4.0",
+    "zone.js": "^0.8.11"
   },
   "devDependencies": {
     "@angular/cli": "^1.0.0",
-    "@angular/compiler-cli": "^4.0.2",
+    "@angular/compiler-cli": "^4.1.3",
     "@types/jasmine": "2.5.38",
     "@types/jquery": "^2.0.37",
     "@types/lodash": "4.14.50",
     "@types/node": "~6.0.60",
     "@types/source-map": "^0.1.28",
     "bootstrap-sass": "^3.3.7",
-    "codelyzer": "~2.0.0",
+    "codelyzer": "~3.0.1",
     "crypto-js": "^3.1.6",
     "del": "^2.2.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
-    "ts-mockito": "^1.2.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "protractor": "~5.1.2",
+    "ts-mockito": "^2.0.0",
+    "ts-node": "~3.0.6",
+    "tslint": "~5.4.3",
+    "typescript": "~2.3.4"
   }
 }
