@@ -16,7 +16,8 @@ module.exports = function (config) {
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
     files: [
-      { pattern: './src/test.ts', watched: false }
+      { pattern: './src/test.ts', watched: false },
+      { pattern: 'node_modules/ngx-bootstrap/bundles/ngx-bootstrap.umd.js', included: false, watched: false }
     ],
     preprocessors: {
       './src/test.ts': ['@angular/cli']
